+++ R/tm_g_bivariate.R
@@ -333,7 +333,6 @@ ui_g_bivariate <- function(id, ...) {
 }
 
 
-#' @importFrom magrittr %>%
 #' @importFrom shinyjs show hide
 srv_g_bivariate <- function(input,
                             output,